Rename components to standard PascalCase & install missing dev dependency
