+++ backend/package.json
@@ -22,5 +22,8 @@
 		"mongodb": "^6.7.0",
 		"mongoose": "^8.4.3",
 		"nodemailer": "^6.9.14"
+	},
+	"devDependencies": {
+		"nodemon": "^3.1.4"
 	}
 }
