+++ frontend/src/components/AnonymousRoute.jsx
